Restructure taxonomy editor files
