+++ src/containers/StructurePage/folderComponents/ConnectFilters.jsx
