+++ src/containers/StructurePage/folderComponents/FilterView.jsx
