+++ src/containers/StructurePage/folderComponents/TopicSettingItems.jsx
@@ -4,8 +4,8 @@ import { Pencil, Plus } from 'ndla-icons/action';
 import { injectT } from 'ndla-i18n';
 import { Filter, DeleteForever } from 'ndla-icons/editor';
 import { Button } from 'ndla-ui';
-import InlineEditField from './InlineEditField';
-import InlineDropdown from './InlineDropdown';
+import InlineEditField from '../../../components/InlineEditField';
+import InlineDropdown from '../../../components/Dropdown/InlineDropdown';
 import ConnectFilters from './ConnectFilters';
 import RoundIcon from '../../../components/RoundIcon';
 import handleError from '../../../util/handleError';
