+++ src/containers/StructurePage/resourceComponents/AddArticleModal.jsx
