+++ src/containers/StructurePage/resourceComponents/Resource.jsx
