+++ src/containers/StructurePage/resourceComponents/ResourceGroup.jsx
