+++ src/containers/StructurePage/resourceComponents/ResourceItems.jsx
