+++ src/containers/StructurePage/resourceComponents/TopicDescription.jsx
